--- src/main/java/com/ardaslegends/repository/UnitTypeRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.ardaslegends.repository;
-
-import com.ardaslegends.domain.UnitType;
-
-import java.util.List;
-
-public interface UnitTypeRepositoryCustom {
-
-    List<UnitType> queryByFactionNames(List<String> factionNames);
-}
