+++ src/main/java/com/ardaslegends/repository/war/army/unit/UnitTypeRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.ardaslegends.repository.war.army.unit;
+
+import com.ardaslegends.domain.UnitType;
+
+import java.util.List;
+
+public interface UnitTypeRepositoryCustom {
+
+    List<UnitType> queryByFactionNames(List<String> factionNames);
+}
