+++ src/main/java/com/ardaslegends/service/UnitTypeService.java
@@ -1,7 +1,7 @@
 package com.ardaslegends.service;
 
 import com.ardaslegends.domain.UnitType;
-import com.ardaslegends.repository.UnitTypeRepository;
+import com.ardaslegends.repository.war.army.unit.UnitTypeRepository;
 import com.ardaslegends.service.exceptions.logic.units.UnitServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
 
 @RequiredArgsConstructor
 @Slf4j
